requirements.txt
bcrypt==4.0.1
python-dotenv==1.0.0
pymongo==4.6.1
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
python-dateutil==2.8.2
